requirements.txt

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
transformers
torch
accelerate
python-dateutil
tzlocal
sentencepiece
python-dotenv